sample_image_converter/requirements.txt
git+https://github.com/Hiroshiba/tornado_instant_webapi
requests
Pillow
